Initial Version
Please fork and improve this.
Things I think should be better:
1) better parsing of my news source (I am running the source with a
friend and we verify everything before posting so no false news).  I
want to see just text not the whole site.  This is currently an RSS
feed.
2) better use of the map.  Maybe we can find a way to use an actual
MKMapView and just parse information onto it?  Also make it easy to
report stuff on the map.